+++ R/funs-string.R
@@ -2,7 +2,7 @@ pl_grepl <- function(pattern, x, fixed = FALSE, ...) {
   if (isTRUE(fixed)) {
     attr(x, "stringr_attr") <- "fixed"
   }
-  pl_str_detect(string = x, pattern = pattern, ...)
+  pl_str_detect_stringr(string = x, pattern = pattern, ...)
 }
 
 pl_paste0 <- function(..., collapse = NULL) {
@@ -14,13 +14,13 @@ pl_paste <- function(..., sep = " ", collapse = NULL) {
   call2(pl$concat_str, !!!clean_dots(...), separator = sep) |> eval_bare()
 }
 
-pl_str_count <- function(string, pattern = "", ...) {
+pl_str_count_stringr <- function(string, pattern = "", ...) {
   check_empty_dots(...)
   pattern <- check_pattern(pattern)
   string$str$count_matches(pattern$pattern, literal = pattern$is_fixed)
 }
 
-pl_str_detect <- function(string, pattern, negate = FALSE, ...) {
+pl_str_detect_stringr <- function(string, pattern, negate = FALSE, ...) {
   check_empty_dots(...)
   pattern <- check_pattern(pattern)
   out <- string$str$contains(pattern$pattern, literal = pattern$is_fixed)
@@ -31,7 +31,7 @@ pl_str_detect <- function(string, pattern, negate = FALSE, ...) {
   }
 }
 
-pl_str_dup <- function(string, times) {
+pl_str_dup_stringr <- function(string, times) {
   if (inherits(times, "RPolarsExpr")) {
     return(string$repeat_by(times)$list$join(""))
   } else if (is.na(times) || times < 0) {
@@ -42,7 +42,7 @@ pl_str_dup <- function(string, times) {
   call2(pl$concat_str, !!!rep(list(string), times)) |> eval_bare()
 }
 
-pl_str_ends <- function(string, pattern, negate = FALSE, ...) {
+pl_str_ends_stringr <- function(string, pattern, negate = FALSE, ...) {
   check_empty_dots(...)
   pattern <- check_pattern(pattern)
 
@@ -61,7 +61,7 @@ pl_str_ends <- function(string, pattern, negate = FALSE, ...) {
 }
 
 # group = 0 means the whole match
-pl_str_extract <- function(string, pattern, group = 0, ...) {
+pl_str_extract_stringr <- function(string, pattern, group = 0, ...) {
   check_empty_dots(...)
   pattern <- check_pattern(pattern)
   # if pattern wasn't passed to pl$col() at this point then it must be parsed
@@ -76,20 +76,20 @@ pl_str_extract <- function(string, pattern, group = 0, ...) {
 # TODO: argument "simplify" should be allowed. It requires the method "unnest"
 # for "struct". When it is implement in r-polars, use this:
 # $list$to_struct()$struct$unnest()
-pl_str_extract_all <- function(string, pattern, ...) {
+pl_str_extract_all_stringr <- function(string, pattern, ...) {
   check_empty_dots(...)
   pattern <- check_pattern(pattern)
   string$str$extract_all(pattern$pattern)
 }
 
-pl_str_length <- function(string, ...) {
+pl_str_length_stringr <- function(string, ...) {
   check_empty_dots(...)
   string$str$len_chars()
 }
 
-pl_nchar <- pl_str_length
+pl_nchar <- pl_str_length_stringr
 
-pl_str_pad <- function(string, width, side = "left", pad = " ", use_width = TRUE, ...) {
+pl_str_pad_stringr <- function(string, width, side = "left", pad = " ", use_width = TRUE, ...) {
   check_empty_dots(...)
   if (isFALSE(use_width)) {
     abort(
@@ -109,19 +109,19 @@ pl_str_pad <- function(string, width, side = "left", pad = " ", use_width = TRUE
   )
 }
 
-pl_str_remove <- function(string, pattern, ...) {
+pl_str_remove_stringr <- function(string, pattern, ...) {
   check_empty_dots(...)
   pattern <- check_pattern(pattern)
   string$str$replace(pattern$pattern, "")
 }
 
-pl_str_remove_all <- function(string, pattern, ...) {
+pl_str_remove_all_stringr <- function(string, pattern, ...) {
   check_empty_dots(...)
   pattern <- check_pattern(pattern)
   string$str$replace_all(pattern$pattern, "")
 }
 
-pl_str_replace <- function(string, pattern, replacement, ...) {
+pl_str_replace_stringr <- function(string, pattern, replacement, ...) {
   check_empty_dots(...)
   pattern <- check_pattern(pattern)
   if (is.character(replacement)) {
@@ -130,7 +130,7 @@ pl_str_replace <- function(string, pattern, replacement, ...) {
   string$str$replace(pattern$pattern, replacement)
 }
 
-pl_str_replace_all <- function(string, pattern, replacement, ...) {
+pl_str_replace_all_stringr <- function(string, pattern, replacement, ...) {
   check_empty_dots(...)
   # named pattern means that names are patterns and values are replacements
   names_pattern <- names(pattern)
@@ -149,12 +149,12 @@ pl_str_replace_all <- function(string, pattern, replacement, ...) {
   out
 }
 
-pl_str_squish <- function(string, ...) {
+pl_str_squish_stringr <- function(string, ...) {
   check_empty_dots(...)
   string$str$replace_all("\\s+", " ")$str$strip_chars()
 }
 
-pl_str_starts <- function(string, pattern, negate = FALSE, ...) {
+pl_str_starts_stringr <- function(string, pattern, negate = FALSE, ...) {
   check_empty_dots(...)
   pattern <- check_pattern(pattern)
 
@@ -172,7 +172,7 @@ pl_str_starts <- function(string, pattern, negate = FALSE, ...) {
   out
 }
 
-pl_str_sub <- function(string, start, end = NULL, ...) {
+pl_str_sub_stringr <- function(string, start, end = NULL, ...) {
   check_empty_dots(...)
   # polars is 0-indexed
   if (start > 0) start <- start - 1
@@ -185,12 +185,12 @@ pl_str_sub <- function(string, start, end = NULL, ...) {
 #
 # Expresses the same objective as me:
 # https://github.com/pola-rs/polars/issues/13649
-pl_str_split <- function(string, pattern, ...) {
+pl_str_split_stringr <- function(string, pattern, ...) {
   check_empty_dots(...)
   string$str$split(by = pattern, inclusive = FALSE)
 }
 
-pl_str_split_i <- function(string, pattern, i, ...) {
+pl_str_split_i_stringr <- function(string, pattern, i, ...) {
   check_empty_dots(...)
   if (i == 0) {
     abort("`i` must not be 0.", call = env_from_dots(...))
@@ -200,26 +200,26 @@ pl_str_split_i <- function(string, pattern, i, ...) {
   string$str$split(by = pattern, inclusive = FALSE)$list$get(i, null_on_oob = TRUE)
 }
 
-pl_str_to_lower <- function(string, ...) {
+pl_str_to_lower_stringr <- function(string, ...) {
   check_empty_dots(...)
   string$str$to_lowercase()
 }
-pl_tolower <- pl_str_to_lower
+pl_tolower <- pl_str_to_lower_stringr
 
-pl_str_to_upper <- function(string, ...) {
+pl_str_to_upper_stringr <- function(string, ...) {
   check_empty_dots(...)
   string$str$to_uppercase()
 }
-pl_toupper <- pl_str_to_upper
+pl_toupper <- pl_str_to_upper_stringr
 
 
-pl_str_to_title <- function(string, ...) {
+pl_str_to_title_stringr <- function(string, ...) {
   check_empty_dots(...)
   string$str$to_titlecase()
 }
-pl_toTitleCase <- pl_str_to_title
+pl_toTitleCase <- pl_str_to_title_stringr
 
-pl_str_trim <- function(string, side = "both", ...) {
+pl_str_trim_stringr <- function(string, side = "both", ...) {
   check_empty_dots(...)
   switch(
     side,
@@ -231,10 +231,10 @@ pl_str_trim <- function(string, side = "both", ...) {
 
 pl_trimws <- function(string, which = "both", ...) {
   check_empty_dots(...)
-  pl_str_trim(string, side = which)
+  pl_str_trim_stringr(string, side = which)
 }
 
-pl_word <- function(string, start = 1L, end = start, sep = " ", ...) {
+pl_word_stringr <- function(string, start = 1L, end = start, sep = " ", ...) {
   check_empty_dots(...)
   string$str$split(sep)$list$gather((start:end) - 1L)$list$join(sep)
 }